Move the tests to a separate PHP namespace
Also add PSR-4 dev-autoloading for this namespace.
